==> notebooks/5 - Cross Validation.ipynb <==
# K-fold cross-validation with 19 datasets

import pandas as pd
import os
import numpy as np
from tqdm import tqdm
from sklearn.model_selection import StratifiedGroupKFold

dataset_root = 'C:/DATASETS/AGE-FER'
dataset_labels_path = os.path.join(dataset_root, '24-datasets.csv')
dataset_imgs_path = os.path.join(dataset_root, 'images')

dtypes = {
    'dataset': 'category',
    'user_id': 'category',
    'name': str,
    'class': 'category',
    'age': 'Int8',
    'gender':'category' ,
    'race': 'category',
    'perspective': 'category',
    'age_group': 'category',
    'subset': 'category',
    'auto_age': bool,
    'auto_gender': bool,
    'age_group_clean': 'category',
    'gaze': 'category',
    'auto_perspective': bool}

output_dir = os.path.join(dataset_root, 'cv-labels')
base_name = '24-datasets'

df_global = pd.read_csv(dataset_labels_path, dtype=dtypes, sep=',', quotechar='"')

# Convert auto_age and auto_gender columns to int
df_global['auto_age'] = df_global['auto_age'].astype(int)
df_global['auto_gender'] = df_global['auto_gender'].astype(int)
df_global['auto_perspective'] = df_global['auto_perspective'].astype(int)

pd.set_option('display.max_rows', 10)
df_global

# Build k stratified subsets of the dataframe
def make_k_subsets(df, k, output_dir, base_name, seed=777):

    # Drop the 'NaN' age group
    df = df.dropna(subset=['age_group_clean'])

    # First create the output directory if it does not exist
    if not os.path.exists(output_dir):
        os.makedirs(output_dir)

    # Use sklearn's StratifiedKFold to split the data
    skf = StratifiedGroupKFold(n_splits=k, shuffle=True, random_state=seed)

    # Add key for groups, of str type. It will be the user_id if it is not in, otherwise it will be the name
    df['key'] = df['user_id'].astype(str)
    df.loc[df['user_id'].isna(), 'key'] = df['name']

    # Add key for strats, of str type. It will be the class + age_group + dataset (same samples of these types in each fold)
    df['strat'] = df['class'].astype(str) + '_' + df['age_group_clean'].astype(str) + '_' + df['dataset'].astype(str)

    # Iterate over the splits
    for i, (train_index, test_index) in enumerate(skf.split(df, df['strat'], groups=df['key'].astype(str))):
        
        # Create the train and test dataframes
        df_train = df.iloc[train_index]
        df_test = df.iloc[test_index]

        # Save the dataframes to CSV files
        train_file = os.path.join(output_dir, base_name + '_train_' + str(i+1) + '.csv')
        test_file = os.path.join(output_dir, base_name + '_test_' + str(i+1) + '.csv')
        df_train.to_csv(train_file, index=False)
        df_test.to_csv(test_file, index=False)

# Make k stratified subsets for the balanced dataframe by age group
make_k_subsets(df_global, k=5, output_dir=output_dir, base_name=base_name, seed=777)

## Make sure all datasets are present in all splits

# Get distinct datasets from df_global
datasets = df_global['dataset'].unique()

### By subset and dataset

df = pd.DataFrame(columns=['subset', 'fold']+list(df_global['dataset'].unique()))
for k in range(1, 6):
    for subset in ['train', 'test']:
        file = os.path.join(output_dir, base_name + '_' + subset + '_' + str(k) + '.csv')
        df_split = pd.read_csv(file, sep=',', quotechar='"')

        # Group by 'class' and 'dataset' and count the occurrences
        grouped_df = df_split.groupby(['dataset']).size().to_frame('count')

        # Add grouped_df to df
        df_row_counts = []
        for dataset in df_global['dataset'].unique():
            if dataset not in grouped_df.index:
                count = 0
            else:
                count = grouped_df.loc[dataset].sum()
            df_row_counts.append(count)
        df.loc[len(df)] = [subset, k] + df_row_counts

pd.set_option('display.max_columns', 30)

df[df['subset'] == 'train']

df[df['subset'] == 'test']

### By subset, dataset and specific class

df = pd.DataFrame(columns=['subset', 'fold', 'class'] + list(df_global['dataset'].unique()))
for k in range(1, 6):
    for subset in ['train', 'test']:
        file = os.path.join(output_dir, base_name + '_' + subset + '_' + str(k) + '.csv')
        df_split = pd.read_csv(file, sep=',', quotechar='"')

        # Group by 'class' and 'dataset' and count the occurrences
        grouped_df = df_split.groupby(['class', 'dataset']).size().to_frame('count')

        # Add grouped_df to df
        for class_label in df_split['class'].unique():
            df_row_counts = []
            for dataset in df_global['dataset'].unique():
                if (class_label, dataset) not in grouped_df.index:
                    count = 0
                else:
                    count = grouped_df.loc[(class_label, dataset)].sum()
                df_row_counts.append(count)
            df.loc[len(df)] = [subset, k, class_label] + df_row_counts

pd.set_option('display.max_columns', 30)

split = 'train'
target_class = 'surprise'
df[(df['subset'] == split) & (df['class'] == target_class)]

### By subset, dataset and specific age group

df = pd.DataFrame(columns=['subset', 'fold', 'age_group_clean'] + list(df_global['dataset'].unique()))
for k in range(1, 6):
    for subset in ['train', 'test']:
        file = os.path.join(output_dir, base_name + '_' + subset + '_' + str(k) + '.csv')
        df_split = pd.read_csv(file, sep=',', quotechar='"')

        # Group by 'age_group_clean' and 'dataset' and count the occurrences
        grouped_df = df_split.groupby(['age_group_clean', 'dataset']).size().to_frame('count')

        # Add grouped_df to df
        for age_group in df_split['age_group_clean'].unique():
            df_row_counts = []
            for dataset in df_global['dataset'].unique():
                if (age_group, dataset) not in grouped_df.index:
                    count = 0
                else:
                    count = grouped_df.loc[(age_group, dataset)].sum()
                df_row_counts.append(count)
            df.loc[len(df)] = [subset, k, age_group] + df_row_counts

pd.set_option('display.max_columns', 30)

split = 'test'
target_group = 'child'
df[(df['subset'] == split) & (df['age_group_clean'] == target_group)]
